nodejs: extract out separate "npm" task
